% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_sorted_structure_plot}
\alias{create_sorted_structure_plot}
\title{Create a Sorted Topic Structure Plot}
\usage{
create_sorted_structure_plot(
  fit,
  samples_metadata,
  grouping_cols,
  colitis_model_name,
  k
)
}
\arguments{
\item{fit}{The \code{fastTopics} fit object.}

\item{samples_metadata}{The metadata data frame, aligned with the fit object.}

\item{grouping_cols}{Character vector of column names for grouping.}

\item{colitis_model_name}{The name of the model, for special conditional formatting.}

\item{k}{The number of topics.}
}
\value{
A \code{ggplot} object.
}
\description{
Generates a structure plot with samples sorted by main topic proportion.
}
