% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_n_ordered_colors}
\alias{select_n_ordered_colors}
\title{Select the First N Ordered Colors
This is an internal helper function, not exported to the user.}
\usage{
select_n_ordered_colors(n, color_palette = DISTINGUISHABLE_COLORS_30)
}
\arguments{
\item{n}{The number of colors to select.}

\item{color_palette}{The palette to choose from.}
}
\value{
A character vector of 'n' hex color codes.
}
\description{
Select the First N Ordered Colors
This is an internal helper function, not exported to the user.
}
